[#28] Rename Character class and its package
`Character` class is renamed to `HangulCharacter` and its package
`character` is renamed to `hangulcharacter`. This renaming is for
avoiding conflicts with `java.lang.Character` class. Package renaming is
for consistency.
