--- app/src/main/java/kr/ac/kaist/team888/character/Mieum.java
@@ -1,5 +1,0 @@
-package kr.ac.kaist.team888.character;
-
-public class Mieum extends Character {
-
-}
