+++ app/src/main/java/kr/ac/kaist/team888/hangulcharacter/Ah.java
@@ -1,0 +1,5 @@
+package kr.ac.kaist.team888.hangulcharacter;
+
+public class Ah extends HangulCharacter {
+
+}
